+++ LazyCache/IAppCache.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Runtime.Caching;
+using System.Threading.Tasks;
 
 namespace LazyCache
 {
     public interface IAppCache
     {
+        ObjectCache ObjectCache { get; }
         void Add<T>(string key, T item);
         void Add<T>(string key, T item, DateTimeOffset absoluteExpiration);
         void Add<T>(string key, T item, TimeSpan slidingExpiration);
@@ -18,5 +20,11 @@ namespace LazyCache
         T GetOrAdd<T>(string key, Func<T> addItemFactory, CacheItemPolicy policy);
 
         void Remove(string key);
+
+        Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory);
+        Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, CacheItemPolicy policy);
+        Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, DateTimeOffset expires);
+        Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, TimeSpan slidingExpiration);
+        Task<T> GetAsync<T>(string key);
     }
 }