+++ LazyCache/Mocks/MockCachingService.cs
@@ -1,8 +1,13 @@
 ﻿using System;
 using System.Runtime.Caching;
+using System.Threading.Tasks;
 
 namespace LazyCache.Mocks
 {
+    /// <summary>
+    ///     A mock implementation IAppCache that does not do any caching.
+    ///     Useful in unit tests or for feature switching to swap in a dependency to disable all caching
+    /// </summary>
     public class MockCachingService : IAppCache
     {
         public void Add<T>(string key, T item)
@@ -32,6 +37,33 @@ namespace LazyCache.Mocks
         {
         }
 
+        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, CacheItemPolicy policy)
+        {
+            return addItemFactory.Invoke();
+        }
+
+        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory)
+        {
+            return addItemFactory.Invoke();
+        }
+
+        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, DateTimeOffset expires)
+        {
+            return addItemFactory.Invoke();
+        }
+
+        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, TimeSpan slidingExpiration)
+        {
+            return addItemFactory.Invoke();
+        }
+
+        public Task<T> GetAsync<T>(string key)
+        {
+            return Task.FromResult(default(T));
+        }
+
+        public ObjectCache ObjectCache => null;
+
 
         public void Add<T>(string key, T item, TimeSpan slidingExpiration)
         {
@@ -49,10 +81,6 @@ namespace LazyCache.Mocks
         public T GetOrAdd<T>(string key, Func<T> addItemFactory, CacheItemPolicy policy)
         {
             return addItemFactory.Invoke();
-        }
-
-        public void Add<T>(string key, DateTime expires, T item)
-        {
         }
     }
 }