+++ Samples/ApiAsyncCachingSample/Controllers/DbQueriesController.cs
@@ -1,0 +1,15 @@
+﻿using System.Web.Http;
+using ApiAsyncCachingSample.Models;
+
+namespace ApiAsyncCachingSample.Controllers
+{
+    public class DbQueriesController : ApiController
+    {
+        [HttpGet]
+        [Route("api/dbQueries")]
+        public int GetDatabaseRequestCounter()
+        {
+            return DbTimeContext.DatabaseRequestCounter();
+        }
+    }
+}