+++ Samples/ApiAsyncCachingSample/Models/DbTime.cs
@@ -1,0 +1,21 @@
+﻿using System;
+
+namespace ApiAsyncCachingSample.Models
+{
+    public class DbTime
+    {
+        public DbTime(DateTime now)
+        {
+            TimeNowInTheDatabase = now;
+        }
+
+        public DbTime()
+        {
+            
+        }
+
+        public DateTime TimeNowInTheDatabase { get; set; }
+
+
+    }
+}